% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allFactors.R
\name{allFactors}
\alias{allFactors}
\title{allFactors}
\source{
http://stackoverflow.com/questions/6424856/r-function-for-returning-all-factors
}
\usage{
allFactors(x)
}
\arguments{
\item{x}{integer}
}
\description{
Function that return all factors of an integer
}
