% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/design.alpha.R
\name{design.alpha}
\alias{design.alpha}
\title{design.alpha}
\source{
agricolae 1.3.3
}
\usage{
design.alpha(
  trt,
  k,
  r,
  serie = 2,
  seed = 0,
  kinds = "Super-Duper",
  randomization = TRUE
)
}
\arguments{
\item{trt}{Treatments}

\item{k}{size block}

\item{r}{Replications}

\item{serie}{number plot, 1: 11,12; 2: 101,102; 3: 1001,1002}

\item{seed}{seed}

\item{kinds}{method for to randomize}

\item{randomization}{TRUE or FALSE - randomize}
}
\description{
Function that generates randomization for for Alpha Lattice Design.
}
