% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/designAlphaLatticeWithDiagCheck.R
\name{designAlphaLatticeWithDiagCheck}
\alias{designAlphaLatticeWithDiagCheck}
\title{Alpha Lattice Design with Diagonal Checks}
\usage{
designAlphaLatticeWithDiagCheck(
  numTrmt,
  numSpatCheck = 1,
  numBlk = 2,
  numRep = 2,
  numTrial = 1,
  numRowPerBlk = 1,
  numRowPerRep = 1,
  numFieldRow = 1,
  serpentine = FALSE,
  topToBottom = TRUE,
  treatName = NULL,
  listName = NULL,
  display = TRUE
)
}
\arguments{
\item{numTrmt}{number of entries to be randomized}

\item{numSpatCheck}{number of spatial checks}

\item{numBlk}{number of blocks}

\item{numRep}{number of replicates}

\item{numTrial}{number of trials (randomization set-ups with the complete set of entries)}

\item{numRowPerBlk}{number of rows per block}

\item{numRowPerRep}{number of rows per replicate}

\item{numFieldRow}{number of field rows}

\item{serpentine}{a logical variable indicating whether plot number will be arranged as serpentine order}

\item{topToBottom}{a logical variable indicating whether plot number will be written from to to bottom}

\item{treatName}{NULL or vector of names associated with the treatment}

\item{listName}{column name associated with the list provided (treatName)}

\item{display}{a logical variable indicating whether randomization parameters will be displayed}
}
\value{
A list containing the following components:
\item{fieldbook}{a data frame}
\item{plan}{a list containing the following components:}
\item{TrmtLayout}{a list whose length is equal to the 
     number of trials containing the treatment layout for each trial} 
\item{PlotNumLayout}{a matrix containing the plot number layout of the experiment} 
\item{RepLayout}{a matrix containing the replication layout of the experiment} 
\item{BlockLayout}{a matrix containing the block layout of the experiment}
}
\description{
Generate randomization and layout for Alpha Lattice Design with Diagonal Checks.
}
\details{
Parameters \code{numFieldRow}, \code{numRowPerRep}  and \code{numRowPerBlk} 
must be specified.  Values of \code{numRowPerBlk} should be a factor of \code{numRowPerRep} while values of 
\code{numRowPerRep} should be a factor of \code{numFieldRow}.
}
\examples{
## Generate randomization and layout for an experiment with 40 entries replicated 2 times.
## Each replicate will have 4 blocks. The experiment will be arrange in a 4 x 20 field.
## Each replicate will be arrange in a 2 x 20, while each block will be arrange in 1 x 10.
## Inserting 1 spatial check in each block diagonally.
alphadc1 <- designAlphaLatticeWithDiagCheck (numTrmt = 24, numSpatCheck = 1,
                                            numBlk = 4, numRep = 2, numTrial = 1,
                                            numRowPerBlk = 1, numRowPerRep = 2,
                                            numFieldRow = 4, serpentine = FALSE, topToBottom = FALSE,
                                            display = TRUE)

}
