% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/designAugmentedAL.R
\name{designAugmentedAL}
\alias{designAugmentedAL}
\title{Augmented Alpha Lattice Design}
\usage{
designAugmentedAL(
  numCheck,
  numTest,
  trmtName = NULL,
  numBlk = 2,
  numRep = 2,
  numTrial = 1,
  genLayout = FALSE,
  numRowPerBlk = 1,
  numRowPerRep = 1,
  numFieldRow = 1,
  serpentine = FALSE,
  topToBottom = TRUE,
  checkTrmtList = NULL,
  testTrmtList = NULL,
  display = TRUE
)
}
\arguments{
\item{numCheck}{number of replicated treatments}

\item{numTest}{number of unreplicated treatments}

\item{trmtName}{NULL or a character string which indicate the name of the treatment to be displayed in the output dataframe}

\item{numBlk}{number of blocks per replicate}

\item{numRep}{number of replicates}

\item{numTrial}{number of trials (randomization set-ups with the complete set of entries)}

\item{genLayout}{logical, whether a layout of the design will be generated or not}

\item{numRowPerBlk}{number of rows per block, if genLayout is TRUE}

\item{numRowPerRep}{number of rows per replicate}

\item{numFieldRow}{number of field rows, if genLayout is TRUE}

\item{serpentine}{a logical variable indicating whether plot number will be arranged as serpentine order}

\item{topToBottom}{a logical variable indicating whether plot number will be written from to to bottom}

\item{checkTrmtList}{NULL or a character vector indicating the names of the replicated treatment}

\item{testTrmtList}{NULL or a character vector indicating the names of the unreplicated treatment}

\item{display}{a logical variable indicating whether randomization parameters will be displayed}
}
\value{
A list containing the following components:
\item{fieldbook}{a data frame}
\item{plan}{a list containing the following components: TrmtLayout and BlockLayout
if \code{genLayout} is \code{FALSE} or a list containing the following components
if \code{genLayout} is \code{TRUE}:}
\item{TrmtLayout}{a data frame containing the treatment layout of the experiment, 
     if \code{genLayout} is \code{FALSE} or a list whose length is equal to the 
     number of trials containing the treatment layout for each trial, 
     if \code{genLayout} is \code{TRUE}} 
\item{PlotNumLayout}{a matrix containing the plot number layout of the experiment} 
\item{RepLayout}{a matrix containing the replication layout of the experiment} 
\item{BlockLayout}{a data frame containing the block layout of the experiment,
     if \code{genLayout} is \code{FALSE} or a matrix containing the block layout of the experiment,
     if \code{genLayout} is \code{TRUE}}
}
\description{
Generate randomization and layout for Augmented Alpha Lattice Design.
}
\details{
If \code{genLayout} is \code{TRUE}, then parameters \code{numFieldRow}, \code{numRowPerRep}  and \code{numRowPerBlk} 
must be specified.  Values of \code{numRowPerBlk} should be a factor of \code{numRowPerRep} while values of 
\code{numRowPerRep} should be a factor of \code{numFieldRow}.

If \code{checkTrmtList} is a character vector, the length should be equal to the \code{numCheck}.
If \code{testTrmtList} is a character vector, the length should be equal to the \code{numTest}.
}
\examples{
## Generate randomization of 8 entries replicated 2 times and 16 unreplicated entries using 
## Augmented Alpha Lattice Design
augAL1 <- designAugmentedAL(numCheck = 8, numTest = 16, numBlk = 4, numRep = 2,  
                            numTrial = 1, genLayout = FALSE)
           
augAL2 <- designAugmentedAL(numCheck = 8, numTest = 16, trmtName = "Variety", numBlk = 4, 
                            numRep = 2, numTrial = 2, genLayout = TRUE, numRowPerBlk = 4, 
                            numRowPerRep = 4, numFieldRow = 8, 
                            serpentine = TRUE, topToBottom = TRUE)
                                                               
}
