% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/designAugmentedRCBD.R
\name{designAugmentedRCB}
\alias{designAugmentedRCB}
\title{Augmented Randomized Complete Block Design}
\usage{
designAugmentedRCB(
  numCheck,
  numTest,
  trmtName = NULL,
  numBlk = 2,
  numTrial = 1,
  genLayout = FALSE,
  numRowPerBlk = 1,
  numFieldRow = 1,
  serpentine = FALSE,
  topToBottom = TRUE,
  checkTrmtList = NULL,
  testTrmtList = NULL,
  display = TRUE
)
}
\arguments{
\item{numCheck}{number of replicated treatments}

\item{numTest}{number of unreplicated treatments}

\item{trmtName}{NULL or a character string which indicate the name of the treatment to be displayed 
in the output dataframe}

\item{numBlk}{number of blocks}

\item{numTrial}{number of trials (randomization set-ups with the complete set of entries)}

\item{genLayout}{logical, whether a layout of the design will be generated or not}

\item{numRowPerBlk}{number of rows per block, if genLayout is TRUE}

\item{numFieldRow}{number of field rows, if genLayout is TRUE}

\item{serpentine}{a logical variable indicating whether plot number will be arranged as serpentine order}

\item{topToBottom}{a logical variable indicating whether plot number will be written from to to bottom}

\item{checkTrmtList}{NULL or a character vector indicating the names of the replicated treatment}

\item{testTrmtList}{NULL or a character vector indicating the names of the unreplicated treatment}

\item{display}{a logical variable indicating whether randomization parameters will be displayed}
}
\value{
A list containing the following components:
\item{fieldbook}{a data frame} 
\item{plan}{a data frame, if \code{genLayout} is \code{FALSE} or a list containing the following components 
if \code{genLayout} is \code{TRUE}:}
\item{TrmtLayout}{a list whose length is equal to the number of trials containing the treatment layout for each trial}
\item{PlotNumLayout}{a matrix containing the plot number information of the experiment}
\item{BlockLayout}{a matrix containing the replication information of the experiment}
}
\description{
Generate randomization and layout for Augmented Randomized Complete
      Block Design.
}
\details{
If \code{genLayout} is \code{TRUE}, then parameters \code{numFieldRow} and \code{numRowPerBlk} must be specified.  
Values of \code{numRowPerBlk} should be a factor of \code{numFieldRow}.

If \code{checkTrmtList} is a character vector, the length should be equal to the \code{numCheck}.
If \code{testTrmtList} is a character vector, the length should be equal to the \code{numTest}.
}
\examples{
## Generate randomization of 6 entries replicated 4 times and 8 unreplicated entries using 
## Augmented Randomized Complete Block Design
augRCBD1 <- designAugmentedRCB(numCheck = 6, numTest = 8, trmtName = "Variety", numBlk = 4,
    numTrial = 2, genLayout = FALSE)
           
augRCBD2 <- designAugmentedRCB(numCheck = 6, numTest = 8, trmtName = "Variety", numBlk = 4,
    numTrial = 2, genLayout = TRUE, numRowPerBlk = 2, numFieldRow = 4, serpentine = TRUE)
                               
}
