% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/designAugmentedRCBDWithDiagCheck.R
\name{designAugmentedRCBWithDiagCheck}
\alias{designAugmentedRCBWithDiagCheck}
\title{Augmented Randomized Complete Block Design with Diagonal Checks}
\usage{
designAugmentedRCBWithDiagCheck(
  numCheck,
  numTest,
  numSpatCheck,
  trmtName = NULL,
  numBlk = 2,
  numTrial = 1,
  numRowPerBlk = 1,
  numFieldRow = 1,
  serpentine = FALSE,
  topToBottom = TRUE,
  checkTrmtList = NULL,
  testTrmtList = NULL,
  spatCheckTrmtList = NULL,
  display = TRUE
)
}
\arguments{
\item{numCheck}{number of replicated treatments}

\item{numTest}{number of unreplicated treatments}

\item{numSpatCheck}{number of spatial checks}

\item{trmtName}{NULL or a character string which indicate the name of the treatment to be displayed 
in the output dataframe}

\item{numBlk}{number of blocks}

\item{numTrial}{number of trials (randomization set-ups with the complete set of entries)}

\item{numRowPerBlk}{number of rows per block, if genLayout is TRUE}

\item{numFieldRow}{number of field rows, if genLayout is TRUE}

\item{serpentine}{a logical variable indicating whether plot number will be arranged as serpentine order}

\item{topToBottom}{a logical variable indicating whether plot number will be written from to to bottom}

\item{checkTrmtList}{NULL or a character vector indicating the names of the replicated treatment}

\item{testTrmtList}{NULL or a character vector indicating the names of the unreplicated treatment}

\item{spatCheckTrmtList}{NULL or a character vector indicating the names of the spatial checks}

\item{display}{a logical variable indicating whether randomization parameters will be displayed}
}
\value{
A list containing the following components:
\item{fieldbook}{a data frame} 
\item{plan}{a list containing the following components:}
\item{TrmtLayout}{a list whose length is equal to the number of trials containing the treatment layout for each trial}
\item{PlotNumLayout}{a matrix containing the plot number information of the experiment}
\item{BlockLayout}{a matrix containing the replication information of the experiment}
}
\description{
Generate randomization and layout for Augmented Randomized Complete
      Block Design.
}
\details{
The parameters \code{numFieldRow} and \code{numRowPerBlk} must be specified.  
Values of \code{numRowPerBlk} should be a factor of \code{numFieldRow}.

If \code{checkTrmtList} is a character vector, the length should be equal to the \code{numCheck}.
If \code{testTrmtList} is a character vector, the length should be equal to the \code{numTest}.
If \code{spatCheckTrmtList} is a character vector, the length should be equal to the \code{numSpatCheck}.
}
\examples{
## Generate randomization of 6 entries replicated 4 times, 8 unreplicated entries and 2 spatial checks
## using Augmented Randomized Complete Block Design with Diagonal checks
                               
}
