% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/designRCBD.R
\name{designRCBD}
\alias{designRCBD}
\title{Randomized Complete Block Design (RCBD)}
\usage{
designRCBD(
  generate,
  numBlk = 2,
  numTrial = 1,
  genLayout = FALSE,
  numFieldRow = 1,
  numRowPerBlk = 1,
  serpentine = FALSE,
  topToBottom = TRUE,
  display = TRUE
)
}
\arguments{
\item{generate}{list of entries to be randomized}

\item{numBlk}{number of replicates or blocks}

\item{numTrial}{number of trials (randomization set-ups with the complete set of entries)}

\item{genLayout}{logical, whether a layout of the design will be generated}

\item{numFieldRow}{number of field rows}

\item{numRowPerBlk}{number of rows per replicate or block}

\item{serpentine}{a logical variable indicating whether plot number will be arranged as serpentine order}

\item{topToBottom}{a logical variable indicating whether plot number will be written from to to bottom}

\item{display}{a logical variable indicating whether randomization parameters will be displayed}
}
\value{
A list containing the following components:
\item{fieldbook}{a data frame}
\item{plan}{a data frame, if \code{genLayout} is \code{FALSE} or a list containing the following components
if \code{genLayout} is \code{TRUE}:}
\item{TrmtLayout}{a list whose length is equal to the number of trials containing the treatment layout for each trial}
\item{PlotNumLayout}{a matrix containing the plot number layout of the experiment}
\item{BlockLayout}{a matrix containing the replication layout of the experiment}
}
\description{
Generate randomization and layout for randomized complete block design (RCBD)
for single factor or factorial experiments.
}
\details{
If \code{genLayout} is \code{TRUE}, then parameters \code{numFieldRow} and \code{numRowPerBlk} must be specified.
Values of \code{numRowPerBlk} should be a factor of \code{numFieldRow}.
}
\examples{
## Generate randomization for an experiment with 10 levels in RCBD replicated 4 times
rcbd1a <- designRCBD(generate = list(Entry = 1:10), numBlk = 4, numTrial = 1, genLayout = FALSE)

## Generate randomization and layout in RCBD for an experiment with l0 levels replicated 4 times in 2 trials
rcbd1b <- designRCBD(generate = list(Entry = 10), numBlk = 4, numTrial = 2, genLayout = TRUE,
   numFieldRow = 10, numRowPerBlk = 5)

## Generate randomization and layout in RCBD for a 5 x 4 factorial experiment replicated 4 times in 2 trials
varietyLevel <- paste("V", 1:5, sep = "")
fertLevel <- paste("F", 1:4, sep = "")
rcbd2a <- designRCBD(generate = list(Variety = varietyLevel, Fertilizer = fertLevel), numBlk = 4,
    numTrial = 2, genLayout = TRUE, numFieldRow = 10, numRowPerBlk = 5, serpentine = TRUE)
}
