% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/designRowColumn.R
\name{designRowColumn}
\alias{designRowColumn}
\title{Row-Column Design}
\usage{
designRowColumn(
  generate,
  numRowBlk = 2,
  numRep = 2,
  numTrial = 1,
  genLayout = FALSE,
  numFieldRow = 2,
  serpentine = FALSE,
  topToBottom = TRUE,
  display = TRUE
)
}
\arguments{
\item{generate}{list of entries to be randomized}

\item{numRowBlk}{number of rows per replicate or number of row blocks per replicate}

\item{numRep}{number of replicates or blocks}

\item{numTrial}{number of trials (randomization set-ups with the complete set of entries)}

\item{genLayout}{logical, whether a layout of the design will be generated or not}

\item{numFieldRow}{number of field rows}

\item{serpentine}{logical, whether plot number will be arranged as serpentine order}

\item{topToBottom}{logical, whether plot number will be written from to to bottom}

\item{display}{a logical variable indicating whether randomization parameters will be displayed}
}
\value{
A list containing the following components:
\item{fieldbook}{a data frame}
\item{plan}{a list containing the following components: TrmtLayout, RowBlockLayout and ColumnBlockLayout, 
if \code{genLayout} is \code{FALSE} or a list containing the following components 
if \code{genLayout} is \code{TRUE}:}
\item{TrmtLayout}{a data frame containing the treatment layout of the experiment, if \code{genLayout} is \code{FALSE} or a list whose length is equal to the 
     number of trials containing the treatment layout for each trial, if \code{genLayout} is \code{TRUE}} 
\item{PlotNumLayout}{a matrix containing the plot number information of the experiment} 
\item{RepLayout}{a matrix containing the replication information of the experiment} 
\item{RowBlockLayout}{a matrix containing the row block information of the experiment}
\item{ColumnBlockLayout}{a matrix containing the column block information of the experiment}
}
\description{
Generate randomization and layout for Row-Column Design.
}
\details{
If \code{genLayout} is \code{TRUE}, then parameters \code{numFieldRow}, \code{numRowBlk}  and \code{numRowPerBlk} 
must be specified.  Values of \code{numRowPerBlk} should be a factor of \code{numRowBlk} while values of 
\code{numRowBlk} should be a factor of \code{numFieldRow}.
}
\examples{
## Generate randomization for an experiment with 20 treatment levels in Row-Column design replicated 4 times.
## Each replicate will have 4 row blocks.
rowcol1 <- designRowColumn(generate = list(Entry = 20), numRowBlk = 4, numRep = 4, numTrial = 1, 
                           genLayout = FALSE)

## Generate randomization and layout for an experiment with 20 treatment levels in Row-Column design replicated 
## 4 times. Each replicate will have 4 row blocks. The experiment will be arrange in a 8 x 10 field while each 
## replicate will be arrange in a 4 x 5.

rowcol2 <- designRowColumn(generate = list(Entry = 20), numRowBlk = 4, numRep = 4, numTrial = 2, 
                           genLayout = TRUE, numFieldRow = 8)

}
