% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DiGGer_fun.R
\name{desPlot}
\alias{desPlot}
\title{desPlot}
\source{
DiGGer 1.0.5
}
\usage{
desPlot(
  dsgn,
  trts = NULL,
  rprop = 1,
  cprop = 1,
  label = TRUE,
  plotchar = NULL,
  plotbndry = TRUE,
  chtdiv = 2,
  bdef = NULL,
  bcol = 1,
  bwd = 2,
  rotate = FALSE,
  new = TRUE,
  cstr = "Range",
  rstr = "Row",
  rlab = TRUE,
  clab = TRUE,
  ...
)
}
\arguments{
\item{dsgn}{A matrix representation of an experimental design.}

\item{trts}{Specific treatment labels to be plotted.}

\item{rprop}{Proportion of the row boundary of an experimental unit to plot.}

\item{cprop}{Proportion of the column boundary of an experimental unit to plot.}

\item{label}{If TRUE print a label for all treatments or specific treatments listed in trts. If FALSE labels are not printed}

\item{plotchar}{A single text character or plotting symbol to be used as the label.}

\item{plotbndry}{If plotbndry is TRUE, a polygon is used to plot the experimental unit.}

\item{chtdiv}{A reduction factor for the plotting symbol.}

\item{bdef}{A matrix rx2. Each row defines the rows and columns in a block template to be plotted.}

\item{bcol}{The colour used for the block boundary.}

\item{bwd}{The width of the block boundary to be plotted.}

\item{rotate}{If TRUE the matrix is rotated 90 degrees for plotting.}

\item{new}{If TRUE a new plot is produced, if FALSE the current plot is added to.}

\item{cstr}{String to label Columns of the matrix.}

\item{rstr}{String to label Rows of the matrix.}

\item{rlab}{If TRUE rows of the design are labelled.}

\item{clab}{If TRUE columns of the design are labelled.}

\item{...}{Other arguments to polygon used for plotting experimental units in the design.}
}
\description{
Plot a matrix representing an experimental design to the current graphics device.
}
