% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diag.layout.R
\name{diag.layout}
\alias{diag.layout}
\title{diag.layout}
\source{
ebsRtools package version 0.2.0
}
\usage{
diag.layout(nCol, nRow, percentChk, plot = FALSE)
}
\arguments{
\item{nCol}{Number of columns in the field, columns are the X coordinates}

\item{nRow}{Number of rows in the field, rows are the Y coordinates}

\item{percentChk}{Percentage of the total plots that should be check plots. Values must be between 0 and 0.5}

\item{plot}{TRUE or FALSE. If should display the field layout in a chart. Default FALSE}
}
\description{
Function that returns a diagonal layout
}
