% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FactorList.R
\name{FactorList}
\alias{FactorList}
\title{Create levels of a factor}
\usage{
FactorList(generate)
}
\arguments{
\item{generate}{a list containing the different levels of treatment combination}
}
\value{
a dataframe.
}
\description{
Generate all treatment combination of several factors.
}
\examples{
FactorList(generate = list(Entry = 1:10))

}
