% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GenerateFactor.R
\name{GenerateFactor}
\alias{GenerateFactor}
\title{Create list of treatments or treatment combinations}
\usage{
GenerateFactor(generate, times = 1)
}
\arguments{
\item{generate}{a list}

\item{times}{number of times the levels will be repeated}
}
\value{
a dataframe.
}
\description{
Generate all treatment combination of several factors.
}
\examples{
GenerateFactor(generate = list(Variety = 4))

}
