% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getMidFloorFactor.R
\name{getMidFloorFactor}
\alias{getMidFloorFactor}
\title{getMidFloorFactor}
\usage{
getMidFloorFactor(x)
}
\arguments{
\item{x}{number}
}
\value{
an integer
}
\description{
Get the lower middle factor of a number
}
\examples{
getMidFloorFactor(20)
}
