% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getRange.R
\name{getRange}
\alias{getRange}
\title{getRange}
\usage{
getRange(
  index,
  topToBottom = TRUE,
  serpentine = FALSE,
  numRowPerGrp,
  numColPerGrp,
  numGrpRow,
  numGrpCol,
  numRow,
  numCol,
  rowIndexLL,
  rowIndexUL,
  colIndexLL,
  colIndexUL
)
}
\arguments{
\item{index}{reference number}

\item{topToBottom}{logical, whether plot number will be written from to to bottom}

\item{serpentine}{logical, whether plot number will be arranged as serpentine order}

\item{numRowPerGrp}{number of rows per block/replicate}

\item{numColPerGrp}{number of rows per block/replicate}

\item{numGrpRow}{number of blocks/replicates per row}

\item{numGrpCol}{number of blocks/replicates per column}

\item{numRow}{number of rows per replicate/field}

\item{numCol}{number of columns per replicate/field}

\item{rowIndexLL}{previous row lower limit}

\item{rowIndexUL}{previous row upper limit}

\item{colIndexLL}{previous column lower limit}

\item{colIndexUL}{previous column upper limit}
}
\value{
a set of integers
}
\description{
Get the next lower and upper limit
}
