% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DiGGer.R
\name{ibDiGGer}
\alias{ibDiGGer}
\title{ibDiGGer}
\source{
DiGGer 1.0.5
}
\usage{
ibDiGGer(
  numberOfTreatments,
  rowsInDesign,
  columnsInDesign,
  rowsInReplicate = NULL,
  columnsInReplicate = NULL,
  rowsInBlock,
  columnsInBlock,
  fixedBlocks = FALSE,
  blockGamma = 1,
  aType = NULL,
  targetAValue = 0,
  maxInterchanges = 1e+05,
  searchIntensity = 100,
  runSearch = TRUE,
  rngSeeds = NULL,
  rngState = NULL,
  treatName = NULL,
  treatNumber = NULL,
  treatRepPerRep = NULL,
  treatGroup = NULL,
  initialDesign = NULL,
  initialSwap = NULL
)
}
\arguments{
\item{numberOfTreatments}{The number of treatments in the design.}

\item{rowsInDesign}{The number of rows in the design.}

\item{columnsInDesign}{The number of columns in the design.}

\item{rowsInReplicate}{The number or rows in the template replicate block.}

\item{columnsInReplicate}{The number of columns in the template replicate block.}

\item{rowsInBlock}{The number of rows in each block.}

\item{columnsInBlock}{The number of columns in each block.}

\item{fixedBlocks}{Logical value whether blocks are treated as fixed.}

\item{blockGamma}{Variance component assigned to blocks.}

\item{aType}{A DiGGer A-measure type: "A++", "Agg", "A22", "A11", "A1+", "Aa2" or "Aaa". The default is "A++" for equally replicated treatments, "Agg" for unequal replication.}

\item{targetAValue}{The search stops in each search phase when the A value is below this value.}

\item{maxInterchanges}{Number of treatment interchanges to test}

\item{searchIntensity}{Percentage of possible interchanges to consider for non-improving interchanges.}

\item{runSearch}{Logical value, whether to run the search immediately.}

\item{rngSeeds}{Seeds c(s1,s2) to control the DiGGer search. s1 must be in the range [0,31328], s2 must be in the range [0,30081].}

\item{rngState}{Current state of the random number generator.}

\item{treatName}{Vector of treatment names to be associated with the design numbers.}

\item{treatNumber}{Vector of treatment numbers to be associated with the design numbers.}

\item{treatRepPerRep}{Vector of replication levels for each treatment within each replicate template block.}

\item{treatGroup}{Vector of group codes (up to 200 distinct values) associated with treatments. Group codes may be used to modify the A-efficiency measure in the optimisation.}

\item{initialDesign}{A matrix rowsInDesign by columnsInDesign giving design numbers in the initial design.}

\item{initialSwap}{A dimension pair or a matrix rowsInDesign by columnsInDesign of swap codes. Only plots with the same swap code may have treatment interchanges during the DiGGer search.}
}
\description{
ibDiGGer modifies a DiGGer object to search for an incomplete block design.
}
