% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layoutAlphaLatticeWithRepCheck.R
\name{layoutAlphaLatticeWithRepCheck}
\alias{layoutAlphaLatticeWithRepCheck}
\title{Layout Creation for Alpha Lattice Design with Repeated Checks}
\usage{
layoutAlphaLatticeWithRepCheck(
  fieldbook,
  trmtLayout,
  numFieldRow,
  numFieldCol,
  numRowPerRep,
  numColPerRep,
  numBlk,
  numRowPerBlk,
  numColPerBlk,
  serpentine = FALSE,
  topToBottom = TRUE
)
}
\arguments{
\item{fieldbook}{a dataframe which is a result from the function \code{randomizeAlphaLatticeWithRepCheck}}

\item{trmtLayout}{a list which is a result from the function \code{randomizeAlphaLatticeWithRepCheck}}

\item{numFieldRow}{number of field rows}

\item{numFieldCol}{number of field columns}

\item{numRowPerRep}{number of rows per replicate}

\item{numColPerRep}{number of columns per replicate}

\item{numBlk}{number of blocks}

\item{numRowPerBlk}{number of rows per block}

\item{numColPerBlk}{number of columns per block}

\item{serpentine}{logical, whether plot number will be arranged as serpentine order}

\item{topToBottom}{logical, whether plot number will be written from to to bottom}
}
\value{
list containing a fieldbook, the layout and plot numbers.
}
\description{
Creates layout for Alpha Lattice Design with Repeated Checks.
}
