% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layoutIBD.R
\name{layoutIBD}
\alias{layoutIBD}
\title{Layout Creation for Incomplete Block Designs}
\usage{
layoutIBD(
  fieldbook,
  trmtLayout,
  numFieldRow,
  numFieldCol,
  numRowPerRep,
  numColPerRep,
  numBlk = NULL,
  numRowPerBlk = NULL,
  numColPerBlk = NULL,
  serpentine = FALSE,
  topToBottom = TRUE
)
}
\arguments{
\item{fieldbook}{a dataframe which is a result from the function \code{randomizeIBD}}

\item{trmtLayout}{a list which is a result from the function \code{randomizeIBD}}

\item{numFieldRow}{number of field rows}

\item{numFieldCol}{number of field columns}

\item{numRowPerRep}{number of rows per replicate}

\item{numColPerRep}{number of columns per replicate}

\item{numBlk}{number of blocks}

\item{numRowPerBlk}{number of rows per block}

\item{numColPerBlk}{number of columns per block}

\item{serpentine}{logical, whether plot number will be arranged as serpentine order}

\item{topToBottom}{logical, whether plot number will be written from to to bottom}
}
\value{
list containing a fieldbook, the layout and plot numbers.
}
\description{
Creates layout Incomplete Block Designs including Alpha Lattice and Row-Column Design.
}
