% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DiGGer.R
\name{prDiGGer}
\alias{prDiGGer}
\title{prDiGGer}
\source{
DiGGer 1.0.5
}
\usage{
prDiGGer(
  numberOfTreatments,
  rowsInDesign,
  columnsInDesign,
  blockSequence = NULL,
  betweenRowCorrModel = "AR",
  betweenRowCorr = 0.5,
  betweenColumnCorrModel = "AR",
  betweenColumnCorr = 0.5,
  treatName = NULL,
  treatNumber = NULL,
  treatRepPerRep = NULL,
  treatGroup = NULL,
  maxInterchanges = 20000,
  targetGroup = 1,
  runSearch = FALSE,
  splitOpt = NULL,
  rngSeeds = NULL
)
}
\arguments{
\item{numberOfTreatments}{Number of distinct treatments in the design.}

\item{rowsInDesign}{Number of rows in the design.}

\item{columnsInDesign}{Number of columns in the design.}

\item{blockSequence}{List of dimension pairs of blocks to be optimised in sequence.}

\item{betweenRowCorrModel}{The correlation pattern between rows may be "AR" AutoRegressive, "MA" Moving Average, or "ID" no correlation.}

\item{betweenRowCorr}{Valid parameter values p, for "AR" -1<p<1, for "MA" -0.5<p<0.5 and for "ID" p=0.}

\item{betweenColumnCorrModel}{As for betweenRowCorrModel.}

\item{betweenColumnCorr}{As for betweenRowCorr.}

\item{treatName}{Vector of texts giving the names of the treatment.}

\item{treatNumber}{Vector of numbers associated with the treatments.}

\item{treatRepPerRep}{Vector of replication levels for each treatment over the whole design.}

\item{treatGroup}{Vector of group codes (up to 200 distinct values) associated with treatments. Group codes may be used to modify the A-efficiency measure in the optimisation.}

\item{maxInterchanges}{Maximum number of interchanges used in the search to spatially optimise the block design created by the call to prDiGGer.}

\item{targetGroup}{Test treatment group used for optimisation in the spatial search phase.}

\item{runSearch}{Logical value, whether to run the final spatial search immediately.}

\item{splitOpt}{Dimension pair used for sub-optimal spatial distribution of replicated entries.}

\item{rngSeeds}{Seeds c(s1,s2) to control the DiGGer search. s1 must be in the range [0,31328], s2 must be in the range [0,30081].}
}
\description{
Generate a DiGGer search for designs with some unreplicated treatments
}
