% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/primeNumber.R
\name{primeNumber}
\alias{primeNumber}
\title{primeNumber}
\source{
https://www.datamentor.io/r-programming/examples/prime-number/
}
\usage{
primeNumber(num)
}
\arguments{
\item{num}{number}
}
\description{
Function that returns 1 if the number is prime else it will return 0.
}
