% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/randALPHA.R
\name{randALPHA}
\alias{randALPHA}
\title{randALPHA}
\source{
ebsRtools 0.2.0
}
\usage{
randALPHA(trt, k, r, tag = 2, nTrial = 1, continue = F)
}
\arguments{
\item{trt}{Number of treatments}

\item{k}{Block size, that is: number of plots per block}

\item{r}{Number of replications (blocks)}

\item{tag}{Format of plot number tag}

\item{nTrial}{Number of occurrences}

\item{continue}{logical}
}
\description{
Function that generates randomization for for Alpha Lattice Design.
}
