% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/randomizeAlphaLatticeEBS.R
\name{randomizeAlphaLatticeEBS}
\alias{randomizeAlphaLatticeEBS}
\title{Randomization for Alpha Lattice design using ebsRTools package}
\usage{
randomizeAlphaLatticeEBS(
  numTrmt,
  numFieldRow,
  numFieldCol,
  numRep,
  numRowPerRep,
  numColPerRep,
  numBlk,
  blksize,
  numRowPerBlk,
  numColPerBlk,
  numBlkRow,
  numBlkCol,
  numRepRow,
  numRepCol,
  topToBottom = NULL,
  serpentine = NULL,
  trmtList = NULL,
  genLayout = FALSE
)
}
\arguments{
\item{numTrmt}{number of treatment}

\item{numFieldRow}{number of field rows}

\item{numFieldCol}{number of field columns}

\item{numRep}{number of replicates}

\item{numRowPerRep}{number of row per replicate}

\item{numColPerRep}{number of column per replicate}

\item{numBlk}{number of blocks per replicate}

\item{blksize}{number of plots per block}

\item{numRowPerBlk}{number of row per block}

\item{numColPerBlk}{number of column per block}

\item{numBlkRow}{number of blocks per row}

\item{numBlkCol}{number of blocks per column}

\item{numRepRow}{number of replicates per row}

\item{numRepCol}{number of replicates per column}

\item{topToBottom}{a logical variable indicating whether plot number will be written from to to bottom}

\item{serpentine}{a logical variable indicating whether plot number will be arranged as serpentine order}

\item{trmtList}{NULL or vector or character containing the levels of the treatment}

\item{genLayout}{logical, whether a layout of the design will be generated}
}
\value{
A list containing dataframe and statistical design array.
}
\description{
Generate randomization.
}
\examples{
randomizeAlphaLatticeEBS (numTrmt = 24, numFieldRow = 8, numFieldCol = 6, 
                          numRep = 2, numRowPerRep = 4, numColPerRep = 6, 
                          numBlk = 6, blksize = 4, numRowPerBlk = 2, numColPerBlk = 2,
                          numBlkRow = 2, numBlkCol = 3, numRepRow = 2, numRepCol = 1,
                          topToBottom = TRUE, serpentine = TRUE, trmtList = paste("Entry",1:24,sep=""),
                          genLayout = FALSE)

}
