% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/randomizeAlphaLatticeWithRepCheck.R
\name{randomizeAlphaLatticeWithRepCheck}
\alias{randomizeAlphaLatticeWithRepCheck}
\title{Randomization for Alpha Lattice designs with Repeated Checks}
\usage{
randomizeAlphaLatticeWithRepCheck(
  numTrmt,
  numFieldRow,
  numFieldCol,
  numRowPerRep,
  numColPerRep,
  numRowPerBlk,
  numColPerBlk,
  trmtRepPerRep,
  trmtGroup,
  trmtList = NULL
)
}
\arguments{
\item{numTrmt}{number of treatment}

\item{numFieldRow}{number of field rows}

\item{numFieldCol}{number of field columns}

\item{numRowPerRep}{number of row per replicate}

\item{numColPerRep}{number of column per replicate}

\item{numRowPerBlk}{number of row per block}

\item{numColPerBlk}{number of column per block}

\item{trmtRepPerRep}{number of replicates of treatment group per replicate}

\item{trmtGroup}{number of treatments per group}

\item{trmtList}{NULL or vector or character containing the levels of the treatment}
}
\value{
A list containing dataframe and statistical design array.
}
\description{
Generate randomization.
}
\examples{
randomizeAlphaLatticeWithRepCheck <- function(numTrmt = 23, numFieldRow = 20, numFieldCol = 4, 
                                             numRowPerRep = 20, numColPerRep = 2, 
                                             numRowPerBlk = 10, numColPerBlk = 1,
                                             trmtRepPerRep = rep(c(1,3,4), c(16,4,3)),
                                             trmtGroup = rep(c(1,2), c(16,7)))

}
