% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/randomizeRCBD.R
\name{randomizeRCBD}
\alias{randomizeRCBD}
\title{Randomization for Randomized Complete Block Design (RCBD)}
\usage{
randomizeRCBD(generate, numBlk = 2, numTrial = 1)
}
\arguments{
\item{generate}{list of entries to be randomized}

\item{numBlk}{number of replicates or blocks}

\item{numTrial}{number of trials (randomization set-ups with the complete set of entries)}
}
\value{
A dataframe.
}
\description{
Generate randomization for randomized complete block design (RCBD)
            for single factor or factorial experiments.
}
\examples{
randomizeRCBD(generate = list(Variety = 4), numBlk = 2, numTrial = 1)

}
