% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/randomizeRCD.R
\name{randomizeRCD}
\alias{randomizeRCD}
\title{Randomization for Row Column design}
\usage{
randomizeRCD(
  numTrmt,
  numFieldRow,
  numFieldCol,
  numRep,
  numRowPerRep,
  numColPerRep,
  trmtList = NULL
)
}
\arguments{
\item{numTrmt}{number of treatment}

\item{numFieldRow}{number of field rows}

\item{numFieldCol}{number of field columns}

\item{numRep}{number of replicates}

\item{numRowPerRep}{number of row per replicate}

\item{numColPerRep}{number of column per replicate}

\item{trmtList}{NULL or vector or character containing the levels of the treatment}
}
\value{
A list containing dataframe and statistical design array.
}
\description{
Generate randomization.
}
\examples{
randomizeRCD(numTrmt = 20, numFieldRow = 16, numFieldCol = 5, 
             numRep = 4, numRowPerRep = 4, numColPerRep = 5) 

}
