% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DiGGer.R
\name{rcDiGGer}
\alias{rcDiGGer}
\title{rcDiGGer}
\source{
DiGGer 1.0.5
}
\usage{
rcDiGGer(
  numberOfTreatments,
  rowsInDesign,
  columnsInDesign,
  rowsInReplicate = NULL,
  columnsInReplicate = NULL,
  twoPhase = TRUE,
  nested = TRUE,
  fixedBlocks = FALSE,
  blockGammas = list(rgamma = 1, cgamma = 1),
  aType = NULL,
  targetAValue = 0,
  maxInterchanges = 1e+05,
  searchIntensity = c(0, 100),
  runSearch = TRUE,
  rngSeeds = NULL,
  rngState = NULL,
  treatName = NULL,
  treatNumber = NULL,
  treatRepPerRep = NULL,
  treatGroup = NULL,
  initialDesign = NULL,
  initialSwap = NULL
)
}
\arguments{
\item{numberOfTreatments}{The number of treatments in the design.}

\item{rowsInDesign}{The number of rows in the design.}

\item{columnsInDesign}{The number of columns in the design.}

\item{rowsInReplicate}{The number or rows in the template replicate block.}

\item{columnsInReplicate}{The number of columns in the template replicate block.}

\item{twoPhase}{If TRUE the bigger block is optimised in the first phase of the search and the smaller in the second phase. The order of the optimisation can be specified as "rowThenCol" or "colThenRow". If FALSE both blocks are optimised in one search phase.}

\item{nested}{If TRUE blocks do not extend past replicates. If a pair of numbers given, the first is the number of rows in a column-block, the second is the number of columns in a row-block. If nested is FALSE, row and column blocks extend across the whole design.}

\item{fixedBlocks}{Logical value whether blocks are treated as fixed.}

\item{blockGammas}{Variance component assigned to random row blocks, rgamma and random column blocks, cgamma.}

\item{aType}{A DiGGer A-measure type: "A++", "Agg", "A22", "A11", "A1+", "Aa2" or "Aaa". The default is "A++" for equally replicated treatments, "Agg" for unequal replication.}

\item{targetAValue}{The search stops in each search phase when the A value is below this value.}

\item{maxInterchanges}{number of treatment interchanges to test in each search phase. If twoPhase = TRUE two numbers may be given.}

\item{searchIntensity}{Percentage of possible interchanges to consider for non-improving interchanges. Two figures may be given for two phase searches.}

\item{runSearch}{Logical value, whether to run the search immediately.}

\item{rngSeeds}{Seeds c(s1,s2) to control the DiGGer search. s1 must be in the range [0,31328], s2 must be in the range [0,30081].}

\item{rngState}{Current state of the random number generator.}

\item{treatName}{Vector of treatment names to be associated with the design numbers.}

\item{treatNumber}{Vector of treatment numbers to be associated with the design numbers.}

\item{treatRepPerRep}{Vector of replication levels for each treatment within each replicate template block.}

\item{treatGroup}{Vector of group codes (up to 200 distinct values) associated with treatments. Group codes may be used to modify the A-efficiency measure in the optimisation.}

\item{initialDesign}{A matrix rowsInDesign by columnsInDesign giving design numbers in the initial design.}

\item{initialSwap}{A dimension pair or a matrix rowsInDesign by columnsInDesign of swap codes. Only plots with the same swap code may have treatment interchanges during the DiGGer search.}
}
\description{
rcDiGGer modifies a DiGGer object to search for a row-column design.
}
