% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/writePlotNumInfo.R
\name{writePlotNumInfo}
\alias{writePlotNumInfo}
\title{writePlotNumInfo}
\usage{
writePlotNumInfo(
  info,
  table,
  tableGrp = NULL,
  numGrpRow,
  numGrpCol,
  numRowPerGrp,
  numColPerGrp,
  incrementWithin = TRUE,
  increment = NULL,
  topToBottom = TRUE
)
}
\arguments{
\item{info}{information to be}

\item{table}{a matrix}

\item{tableGrp}{NULL or a matrix}

\item{numGrpRow}{a numerical value}

\item{numGrpCol}{a numerical value}

\item{numRowPerGrp}{a numerical value}

\item{numColPerGrp}{a numerical value}

\item{incrementWithin}{a logical value}

\item{increment}{NULL or a numerical value}

\item{topToBottom}{logical, whether plot number will be written from to to bottom}
}
\value{
a matrix
}
\description{
Utility function.
}
